New way of running the simulation test.
Sahit, please try running ` roslaunch buzzmobile simulation_test.launch`
locally and seeing what happens. I get errors that the controller
manager cannot find dependencies, and I want to see if that's true for
you as well.

I think this implies some issues in either the urdf or in
sim_car_interface.cpp, but I'm not 100% sure of that.